Add utils module and refactored util functions
